Add files via upload
Fix codes in JavaScript and CSS. Fix logo size, main-content which is in the center of the screen. Fix site on phones, etc. Fixed a bug with the last question (if the player did not select a question, the system transfers the player to a blank sheet, it should have been on the Score board), now everything works correctly