Add files via upload
Changed colors and design of navigation buttons.
Text font.
Optimizing the site for all devices, including phones, which I had a big problem with, optimized the size of the content that is in the middle, as well as the logo covering most of it or getting in the way. Ability to scroll below for devices with low resolution to view all content on the page.